# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for simpo pharmacy management system supporting POS, inventory, reporting, and multi-branch operations.",
        "title": "simpo Pharmacy Management System API",
        "termsOfService": "https://simpo.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@simpo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/whitelist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all email domains in the whitelist (SYSTEM_ADMIN only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "whitelist"
                ],
                "summary": "List all whitelisted domains",
                "responses": {
                    "200": {
                        "description": "List of whitelist entries",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/whitelist.WhitelistEntryResponse"
                                            }
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden - SYSTEM_ADMIN only",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new email domain to the whitelist for staff self-registration (SYSTEM_ADMIN only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "whitelist"
                ],
                "summary": "Add email domain to whitelist",
                "parameters": [
                    {
                        "description": "Add whitelist entry request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whitelist.AddWhitelistEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Whitelist entry created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/whitelist.WhitelistEntryResponse"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden - SYSTEM_ADMIN only",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Domain already exists",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errors.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errors.ErrorInfo"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errors.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Additional fields for internal use",
                    "type": "string"
                },
                "detail": {
                    "description": "Human-readable explanation of this occurrence",
                    "type": "string"
                },
                "details": {},
                "instance": {
                    "description": "URI reference identifying this specific occurrence",
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "retry_after": {
                    "type": "integer"
                },
                "status": {
                    "description": "HTTP status code (required by RFC 7807)",
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "title": {
                    "description": "Short, human-readable summary",
                    "type": "string"
                },
                "type": {
                    "description": "RFC 7807 required fields",
                    "type": "string"
                }
            }
        },
        "errors.Links": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                },
                "self": {
                    "type": "string"
                }
            }
        },
        "errors.Meta": {
            "type": "object",
            "properties": {
                "links": {
                    "$ref": "#/definitions/errors.Links"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "errors.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/errors.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/errors.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "whitelist.AddWhitelistEntryRequest": {
            "type": "object",
            "required": [
                "default_role",
                "domain"
            ],
            "properties": {
                "default_role": {
                    "description": "DefaultRole is the default role assigned to users registering with this domain\nRequired: true\nEnum: SYSTEM_ADMIN, OWNER, CASHIER\nExample: \"CASHIER\"",
                    "type": "string",
                    "enum": [
                        "SYSTEM_ADMIN",
                        "OWNER",
                        "CASHIER"
                    ],
                    "example": "CASHIER"
                },
                "description": {
                    "description": "Description is an optional description of the whitelist entry\nRequired: false\nExample: \"Simpo Pharmacy staff domain\"",
                    "type": "string",
                    "example": "Simpo Pharmacy staff domain"
                },
                "domain": {
                    "description": "Domain is the email domain to whitelist (e.g., \"simpo.pharmacy\")\nRequired: true\nExample: \"simpo.pharmacy\"",
                    "type": "string",
                    "example": "simpo.pharmacy"
                }
            }
        },
...
        "whitelist.WhitelistEntryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt is the timestamp when the entry was created\nExample: \"2026-05-12T00:00:00Z\"",
                    "type": "string",
                    "example": "2026-05-12T00:00:00Z"
                },
                "default_role": {
                    "description": "DefaultRole is the default role for this domain\nExample: \"CASHIER\"",
                    "type": "string",
                    "example": "CASHIER"
                },
                "description": {
                    "description": "Description is the description of the whitelist entry\nExample: \"Simpo Pharmacy staff domain\"",
                    "type": "string",
                    "example": "Simpo Pharmacy staff domain"
                },
                "domain": {
                    "description": "Domain is the whitelisted email domain\nExample: \"simpo.pharmacy\"",
                    "type": "string",
                    "example": "simpo.pharmacy"
                },
                "id": {
                    "description": "ID is the unique identifier for the whitelist entry\nExample: 1",
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "description": "UpdatedAt is the timestamp when the entry was last updated\nExample: \"2026-05-12T00:00:00Z\"",
                    "type": "string",
                    "example": "2026-05-12T00:00:00Z"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}